+++ src/lib/pages/publish-module/publish.tsx
@@ -19,6 +19,7 @@ import { ErrorMessageRender } from "lib/components/ErrorMessageRender";
 import { EstimatedFeeRender } from "lib/components/EstimatedFeeRender";
 import { CustomIcon } from "lib/components/icon";
 import PageContainer from "lib/components/PageContainer";
+import { CelatoneSeo } from "lib/components/Seo";
 import { useTxBroadcast } from "lib/hooks";
 import type { DecodeModuleQueryResponse } from "lib/services/types";
 import type { Option } from "lib/types";
@@ -208,6 +209,7 @@ export const PublishModule = ({
 
   return (
     <>
+      <CelatoneSeo pageName="Publish / Republish Modules" />
       <PageContainer p={0}>
         <Box minH="inherit" maxW="1440px" mx="auto">
           <Grid